renamed fileSymbolsTag.ts -> symbolsTag.ts
