+++ lib/main/atom/views/symbols/symbolsTag.ts
